requirements.txt

feedparser
python-dateutil
beautifulsoup4
polars
sqlite-utils
feedgen
